projects/multi-agent-trading-rl/requirements.txt
langgraph
langchain
langchain-openai
pinecone-client
pandas
numpy
pydantic
python-dotenv
yfinance
ta
gymnasium
stable-baselines3
matplotlib
scikit-learn
